+++ apps/mobile/src/screens/Transaction/components/Actions/UnknownAction.tsx
@@ -45,11 +45,6 @@ export const UnknownAction: React.FC<Props> = ({
   const { styles } = useTheme2024({ getStyle });
   const { t } = useTranslation();
 
-  console.log('UnknownAction render', {
-    data,
-    chain,
-  });
-
   return (
     <>
       <ScrollView
